implemented discussed target file concept
