+++ ModTekUnitTests/ModTekUnitTests.csproj
@@ -52,7 +52,7 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="JSONPathMergerTests.cs" />
+    <Compile Include="AdvancedJSONMergeInstructionTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
